GH-1374: fix NPE on server.destroy() when queue mode is disabled

+++ leshan-server-cf/src/main/java/org/eclipse/leshan/server/californium/LeshanServer.java
@@ -430,7 +430,9 @@ public class LeshanServer {
             ((Stoppable) requestSender).stop();
         }
 
-        presenceService.destroy();
+        if (presenceService != null) {
+            presenceService.destroy();
+        }
 
         LOG.info("LWM2M server destroyed.");
     }
